python-worker/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asyncio==4.0.0
beautifulsoup4==4.14.2
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.10.0
google-auth==2.43.0
google-genai==1.52.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
http
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.2
markdownify==0.13.1
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6
nexus-rpc==1.1.0
numpy==1.26.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
protobuf==6.33.1
psycopg2==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jobspy==1.1.82
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
regex==2024.11.6
requests==2.32.5
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sympy==1.14.0
temporalio==1.19.0
tenacity==9.1.2
threadpoolctl==3.6.0
tls-client==1.0.1
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.1
triton==3.5.1
types-protobuf==6.32.1.20251105
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
